pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]
name = "nomad-unisyscat"
description = "A example plugin for a demonstration for UniSysCat."
version = "0.1.0"
readme = "README.rst"
requires-python = ">=3.9"
authors = [
    { name = "Pepe Marquez", email = "jose.marquez@physik.hu-berlin.de" },
]
maintainers = [
    { name = "Pepe Marquez", email = "jose.marquez@physik.hu-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = ["nomad-lab>=1.2.2dev578"]

[project.urls]
Repository = "https://github.com/Pepe-Marquez/nomad-unisyscat"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
myparser = "nomad_unisyscat.parsers:myparser"
mypackage = "nomad_unisyscat.schema_packages:mypackage"

myapp = "nomad_unisyscat.apps:myapp"
